feat(PaginatedMessage): made it possible to add custom `link` buttons with `setActions` / `addActions` / `addAction`
This makes it possible to add `LINK` style buttons to the paginated message's actions.
Previously it was set up to lock out this possibility.
I probably had some issues implementing it in a sensible way and gave up, I don't remember.

I also made it possible to set a second optional parameter of `setActions` to true
to have the library merge in `PaginatedMessage.defaultActions`, something that is a
common use case when using this particular function.
This change will save users from having to manually add the default actions.

I'm not sure if this change should be considered `semver:major`. @vladfrangu what do you say?
I'm primarily asking because `PaginatedMessageAction` was changed from an `interface` to a `type`
which subsequently caused this issue at least in Skyra's code:

![](https://media.discordapp.net/attachments/759750007087431721/931976453879373864/unknown.png?width=1440&height=223)

Easily fixed of course by just changing it to a `type`, but the code did not compile.
Is that `semver:major`?

```typescript
type HelpPaginatedMessageAction = PaginatedMessageAction & {
	selectMenuIndex?: 'set-1' | 'set-2';
};
```