feat(PaginatedMessage): ensure it can run on global chat input commands
This also introduces a new potential emitted warning when the client is not
properly configured to handle DM commands.

Whether the warning is emitted at all can be configured so people can choose to not get it at all.
It will also only be emitted once per instance of PaginatedMessage.

On the note of the warning being per instance of PaginatedMessage, yes that does often mean that
people will get it for every time that `run` is called because it's common (and recommended)
to call `new PaginatedMessage` inside of the `*Run` method.
I don't think it's fitting to keep track of that boolean outside of the class though.

Calling run twice example:
![](https://favna.s-ul.eu/SOJzv2G9.png)