refactor: rewrite `/result`
This update brings a simplification to the `result` package, removing the duplication of the `Option` and `Result` types and their respective implementations. With this new approach, the `Option` and `Result` types now have a single implementation, which is more in line with the Rust language.

BREAKING CHANGE: Renamed `Result.is` to `Result.isResult`
BREAKING CHANGE: Renamed `Option.is` to `Option.isOption`